requirements.txt
tensorboard==2.16.2
tqdm==4.66.4
jiwer==3.0.4
transformers==4.41.1
streamlit==1.35.0
PyYAML==6.0.1
numpy==1.26.4
matplotlib==3.9.0
